Week 7 changes
The text sizing media queries for larger screens.
Text alignment of the copyright and the navigation in the footer.
Fixed sizing around footer.
Extra layout for the data lists on large screens.
The new content at the bottom of the moons page.
Fixed extension of footer in moons and exploration page.